data-processing-agent-remove/backend/requirements_supervisor.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
boto3>=1.35.0
requests>=2.31.0
pydantic>=2.5.0
uv
strands-agents
strands-agents-tools
bedrock-agentcore
bedrock-agentcore-starter-toolkit
mcp